Cosine_version/requirements.txt
flask==3.0.0
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
faiss-cpu==1.7.4
numpy==1.24.3
werkzeug==3.0.1
